+++ src/Shared/Models/PVDay.cs
@@ -6,7 +6,7 @@ public class PVDay
 {
     public int DayNumber { get; set; }
     public string DayName { get; set; }
-    public decimal TotalKiloWattHours { get; set; }
+    public double TotalKiloWattHours { get; set; }
     public IEnumerable<WattPeak> WattPeaks { get; set; }
     public int GraphTickCount { get; set; }
     public int[] GraphRange { get; set; }
